+++ integrations/google_ai/tests/generators/chat/test_chat_gemini.py
@@ -117,7 +117,7 @@ class TestGoogleAIGeminiChatGenerator:
         with patch("haystack_integrations.components.generators.google_ai.chat.gemini.genai.configure"):
             gemini = GoogleAIGeminiChatGenerator()
         assert gemini.to_dict() == {
-            "type": "haystack_integrations.components.generators.google_ai.chat.gemini.GoogleAIGeminiChatGenerator",
+            "type": TYPE,
             "init_parameters": {
                 "api_key": {"env_vars": ["GOOGLE_API_KEY"], "strict": True, "type": "env_var"},
                 "model": "gemini-1.5-flash",
@@ -157,7 +157,7 @@ class TestGoogleAIGeminiChatGenerator:
                 tools=tools,
                 tool_config=tool_config,
             )
-        assert gemini.to_dict() == {
+        expected_dict = {
             "type": TYPE,
             "init_parameters": {
                 "api_key": {"env_vars": ["GOOGLE_API_KEY"], "strict": True, "type": "env_var"},
@@ -191,6 +191,17 @@ class TestGoogleAIGeminiChatGenerator:
                 },
             },
         }
+
+        # add outputs_to_string, inputs_from_state and outputs_to_state tool parameters for compatibility with
+        # haystack-ai>=2.12.0
+        if hasattr(tools[0], "outputs_to_string"):
+            expected_dict["init_parameters"]["tools"][0]["data"]["outputs_to_string"] = tools[0].outputs_to_string
+        if hasattr(tools[0], "inputs_from_state"):
+            expected_dict["init_parameters"]["tools"][0]["data"]["inputs_from_state"] = tools[0].inputs_from_state
+        if hasattr(tools[0], "outputs_to_state"):
+            expected_dict["init_parameters"]["tools"][0]["data"]["outputs_to_state"] = tools[0].outputs_to_state
+
+        assert gemini.to_dict() == expected_dict
 
     def test_from_dict(self, monkeypatch):
         monkeypatch.setenv("GOOGLE_API_KEY", "test")
@@ -324,6 +335,21 @@ class TestGoogleAIGeminiChatGenerator:
 
         if not hasattr(pipeline, "_connection_type_validation"):
             expected_dict.pop("connection_type_validation")
+
+        # add outputs_to_string, inputs_from_state and outputs_to_state tool parameters for compatibility with
+        # haystack-ai>=2.12.0
+        if hasattr(tool, "outputs_to_string"):
+            expected_dict["components"]["generator"]["init_parameters"]["tools"][0]["data"][
+                "outputs_to_string"
+            ] = tool.outputs_to_string
+        if hasattr(tool, "inputs_from_state"):
+            expected_dict["components"]["generator"]["init_parameters"]["tools"][0]["data"][
+                "inputs_from_state"
+            ] = tool.inputs_from_state
+        if hasattr(tool, "outputs_to_state"):
+            expected_dict["components"]["generator"]["init_parameters"]["tools"][0]["data"][
+                "outputs_to_state"
+            ] = tool.outputs_to_state
 
         assert pipeline_dict == expected_dict