+++ integrations/mistral/tests/test_mistral_document_embedder.py
@@ -68,6 +68,7 @@ class TestMistralDocumentEmbedder:
                 "progress_bar": True,
                 "meta_fields_to_embed": [],
                 "embedding_separator": "\n",
+                "http_client_kwargs": None,
             },
         }
 
@@ -99,6 +100,7 @@ class TestMistralDocumentEmbedder:
                 "progress_bar": False,
                 "meta_fields_to_embed": ["test_field"],
                 "embedding_separator": "-",
+                "http_client_kwargs": None,
             },
         }