+++ integrations/stackit/tests/test_stackit_text_embedder.py
@@ -48,6 +48,7 @@ class TestSTACKITTextEmbedder:
                 "organization": None,
                 "prefix": "",
                 "suffix": "",
+                "http_client_kwargs": None,
             },
         }
 
@@ -71,6 +72,7 @@ class TestSTACKITTextEmbedder:
                 "organization": None,
                 "prefix": "START",
                 "suffix": "END",
+                "http_client_kwargs": None,
             },
         }